src/handlers/CreateJpeg/requirements.txt
# This must be specified from the root of the project for SAM
-e src/common/

aws_lambda_powertools
boto3
mypy_boto3_s3
pillow
rawpy
